feat: add docs + examples
