+++ lerna.json
@@ -28,5 +28,5 @@
       "registry": "https://registry.npmjs.org/"
     }
   },
-  "version": "0.19.0"
+  "version": "0.19.3"
 }
