+++ ui/select/src/SelectValue.tsx
@@ -1,15 +1,16 @@
-import * as React from "react";
-import { SelectContext } from "./SelectRoot";
+import React, { useEffect } from "react";
+import { SelectContext } from "./Select";
 
 import { theme, styled } from "@washingtonpost/wpds-theme";
 import { unstyledInputStyles } from "@washingtonpost/wpds-input-shared";
 
 import * as SelectPrimitive from "@radix-ui/react-select";
 import type * as WPDS from "@washingtonpost/wpds-theme";
+import { SelectValueProps as RadixSelectValueProps } from "@radix-ui/react-select";
 
 const StyledValue = styled("div", {
   ...unstyledInputStyles,
-  width: "90%",
+  width: "100%",
   textAlign: "start",
   cursor: "pointer",
   height: theme.space[300],
@@ -26,16 +27,28 @@ const StyledValue = styled("div", {
   },
 });
 
-export type SelectValueProps = {
-  /** Used to insert select elements into the root component*/
-  children?: React.ReactNode;
-  /** Overrides for the input text styles. Padding overrides affect the input container and  */
-  css?: WPDS.CSS;
-} & React.ComponentPropsWithRef<typeof StyledValue>;
+type SelectValueVariants = WPDS.VariantProps<typeof StyledValue>;
+
+type SelectValueProps = SelectValueVariants &
+  RadixSelectValueProps & {
+    /** Used to insert select elements into the root component*/
+    children?: React.ReactNode;
+    /** Overrides for the input text styles. Padding overrides affect the input container and  */
+    css?: WPDS.CSS;
+    // /** Used to specify placeholder text. Can also be passed in as a child */
+    placeholder?: string | undefined;
+  };
 
 export const SelectValue = React.forwardRef<HTMLDivElement, SelectValueProps>(
   ({ children, ...props }: SelectValueProps, ref) => {
-    const { currentValue, disabled } = React.useContext(SelectContext);
+    const { currentValue, disabled, setIsFloating } =
+      React.useContext(SelectContext);
+
+    useEffect(() => {
+      if (children) {
+        setIsFloating(true);
+      }
+    }, [children, setIsFloating]);
 
     return (
       <StyledValue {...props} isDisabled={disabled}>
