+++ ui/select/src/index.ts
@@ -3,6 +3,5 @@ export * from "./SelectContent";
 export * from "./SelectGroup";
 export * from "./SelectItem";
 export * from "./SelectLabel";
-export * from "./SelectRoot";
 export * from "./SelectTrigger";
 export * from "./SelectValue";
